+++ bpf-gpl/arp_types.h
@@ -1,0 +1,31 @@
+// Project Calico BPF dataplane programs.
+// Copyright (c) 2021 Tigera, Inc. All rights reserved.
+//
+// This program is free software; you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation; either version 2 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License along
+// with this program; if not, write to the Free Software Foundation, Inc.,
+// 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+#ifndef __CALI_ARP_TYPES_H__
+#define __CALI_ARP_TYPES_H__
+
+struct arp_key {
+	__u32 ip;
+	__u32 ifindex;
+};
+
+struct arp_value {
+	char mac_src[6];
+	char mac_dst[6];
+};
+
+#endif /* __CALI_ARP_TYPES_H__ */
