+++ bpf-gpl/fib.h
@@ -70,32 +70,56 @@ static CALI_BPF_INLINE int forward_or_drop(struct cali_tc_ctx *ctx)
 		CALI_DEBUG("Redirect to the same interface (%d) failed.\n", ctx->skb->ifindex);
 		goto deny;
 	} else if (rc == CALI_RES_REDIR_IFINDEX) {
-		struct arp_value *arpv;
 		__u32 iface = state->ct_result.ifindex_fwd;
 
-		struct arp_key arpk = {
-			.ip = state->ip_dst,
-			.ifindex = iface,
-		};
+		if (CALI_F_FROM_WEP) {
+			struct arp_value *arpv;
+			struct arp_key arpk = {
+				.ip = state->ip_dst,
+				.ifindex = iface,
+			};
 
-		arpv = cali_v4_arp_lookup_elem(&arpk);
-		if (!arpv) {
-			CALI_DEBUG("ARP lookup failed for %x dev %d\n",
-					bpf_ntohl(state->ip_dst), iface);
-			goto skip_redir_ifindex;
+
+			arpv = cali_v4_arp_lookup_elem(&arpk);
+			if (!arpv) {
+				CALI_DEBUG("ARP lookup failed for %x dev %d\n",
+						bpf_ntohl(state->ip_dst), iface);
+				goto skip_redir_ifindex;
+			}
+
+			/* Revalidate the access to the packet */
+			if (skb_refresh_validate_ptrs(ctx, UDP_SIZE)) {
+				ctx->fwd.reason = CALI_REASON_SHORT;
+				CALI_DEBUG("Too short\n");
+				goto deny;
+			}
+
+			/* Patch in the MAC addresses that should be set on the next hop. */
+			struct ethhdr *eth_hdr = ctx->data_start;
+			__builtin_memcpy(&eth_hdr->h_dest, arpv->mac_dst, ETH_ALEN);
+			__builtin_memcpy(&eth_hdr->h_source, arpv->mac_src, ETH_ALEN);
+		} else if (CALI_F_FROM_HEP && ctx->state->tun_ip == 0) {
+			/* We received a paket from HEP, we marked it for direct forward, we need to
+			 * record ARP for the return path.
+			 */
+
+			/* Revalidate the access to the packet */
+			if (skb_refresh_validate_ptrs(ctx, UDP_SIZE)) {
+				ctx->fwd.reason = CALI_REASON_SHORT;
+				CALI_DEBUG("Too short\n");
+				goto deny;
+			}
+
+			/* XXX if only verifier would allow us.
+			 * arp_record_reverse(ctx);
+			 */
+			struct arp_key arpk = {
+				.ip = state->ip_src,
+				.ifindex = ctx->skb->ifindex,
+			};
+
+			cali_v4_arp_update_elem(&arpk, ctx->eth, 0);
 		}
-
-		/* Revalidate the access to the packet */
-		if (skb_refresh_validate_ptrs(ctx, UDP_SIZE)) {
-			ctx->fwd.reason = CALI_REASON_SHORT;
-			CALI_DEBUG("Too short\n");
-			goto deny;
-		}
-
-		/* Patch in the MAC addresses that should be set on the next hop. */
-		struct ethhdr *eth_hdr = ctx->data_start;
-		__builtin_memcpy(&eth_hdr->h_dest, arpv->mac_dst, ETH_ALEN);
-		__builtin_memcpy(&eth_hdr->h_source, arpv->mac_src, ETH_ALEN);
 
 		rc = bpf_redirect(iface, 0);
 		if (rc == TC_ACT_REDIRECT) {
